Refactor JavaScript
Implemented ES Modules to de-duplicate all the different functions that are repeated across pages
