+++ frontend/static/js/utils/dragUtils.js
@@ -1,0 +1,6 @@
+export function enableDragAndDrop(container, onUpdateCallback) {
+    new Sortable(container, {
+        animation: 150,
+        onEnd: onUpdateCallback
+    });
+}
