+++ frontend/templates/pdf-merge.html
@@ -46,5 +46,5 @@
 {% block post_scripts %}
 <script src="https://cdnjs.cloudflare.com/ajax/libs/Sortable/1.14.0/Sortable.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/pdf.js/2.10.377/pdf.min.js"></script>
-<script src="{{ url_for('static', filename='js/pdf-merge.js') }}"></script>
+<script type="module" src="{{ url_for('static', filename='js/pdf-merge.js') }}"></script>
 {% endblock %}
