+++ frontend/templates/pdf-organize.html
@@ -42,5 +42,5 @@
 {% block post_scripts %}
 <script src="https://cdnjs.cloudflare.com/ajax/libs/Sortable/1.14.0/Sortable.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/pdf.js/2.10.377/pdf.min.js"></script>
-<script src="../static/js/pdf-organize.js"></script>
+<script type="module" src="../static/js/pdf-organize.js"></script>
 {% endblock %}
