+++ frontend/templates/pdf-split.html
@@ -62,5 +62,5 @@
 
 {% block post_scripts %}
 <script src="https://cdnjs.cloudflare.com/ajax/libs/pdf.js/2.10.377/pdf.min.js"></script>
-<script src="../static/js/pdf-split.js"></script>
+<script type="module" src="../static/js/pdf-split.js"></script>
 {% endblock %}
